Add LXD support for ceph-access relation
The nova-lxd driver is growing support for ceph storage backends,
but as this is a host based integration, it relies on keyrings
being placed in /etc/ceph for cephx authentication, rather than
using some other secret storage mechanism as done for libvirt.

Tweak the ceph-access-relation-changed handling to deal with this
and ensure that the ceph-common package is installed with relations
of this type are joined.

Change-Id: I887e0be007c5606614e00c6a374416368d675c4d