my_ip should belong to internal-network rather than data-network
When defining os-data-network=10.10.1.0/24,
os-internal-network=10.12.1.0/24 and not defining os-data-space,
private-address of nova-compute units will be in 10.10.1.0/24 because
now 10.10.1.0/24 and 10.12.1.0/24 are all in internal-network.

nova-cloud-controller units can't collect public keys of nova-compute
units because they don't have NIC belonged to 10.10.1.0/24. This will
lead to some problems when resizing/migrating VM.

This problem can be fixed by defining os-internal-space for
cloud-compute because at this time private-address of nova-compute
units will be in 10.12.1.0/24.

    bindings:
      cloud-compute: *os-internal-space

However, the nova-compute charm has an os-internal-network option,
this should also be referenced as a way to do this since not
everybody uses spaces. This patch is used to deliver this point,
then we can fix this problem by running the following command:

juju config nova-compute os-internal-network='10.12.1.0/24'

Change-Id: I70b096fd709e5ec8d8a47d2d11f71ed3ea780c5d
Closes-Bug: 1686882