Enable support for auto-converge and post-copy
Implement missing configuration options and context to support
configuration of:

  live_migration_permit_auto_converge
  live_migration_permit_post_copy

in nova.conf. The configuration parameters are only added to the
nova.conf template when live-migration is actually enabled, since they
have no purpose without live-migration enabled. The options are added to
version Newton and upwards as that is the Nova version since they were
first supported.

Change-Id: I1bb3cee4ac532d0867b4297c742707668566a527
Closes-Bug: #1799916