Enable bionic/stein and disco/stein functional tests
Drop service check for ``glance-registry``.  The service is
removed as of Stein, and it really is not this charm functional
tests job of checking this anyway.  Any such detail should
be deferred to the ``charm-glance`` functional tests.

Update templates/usr.bin.nova-compute as nova-compute requires
read access to /etc/mime.types.

Change-Id: I9a894ef2a3395ead2509edd4a6b8885a1739e974