Start libvirt in post-series-upgrade
A recent change (commit ceab1e91dc2e3948f6ba7c121c1801ad1641643c)
removed libvirt from the pause/resume list of services. This affected
series upgrade. Libvirt stayed down and did not start back up on resume.

This change checks the hook being executed and if it is
post-series-upgrade it includes libvirt as a service to start on resume.

Closes-Bug: #1839495
Change-Id: Ibfa24b678c1077b441464da8c38114ce23d14963