Ensure python3-hvac is installed for charms with encypt option
The referenced bug is essentially: make vault:secrets relation to vault
but keep the 'encrypt' option as False. In this case, the Context
handling code in charm-helpers is expecting python3-hvac to be
available, but it is only installed if the encrypt option is set to
True.  Hence the charm crashes.  This resolves that crash.

Note the related charm-helpers fix [1].

[1]: https://github.com/juju/charm-helpers/pull/431

Change-Id: Ic02d4d4d3c3b423fa28cd171b126ed4a444fc646
Closes-bug: #1862085