Fix oslo notification format configuration option
This patch removes the hardcoded notification_format value in nova.conf
in order to fill it with the configuration option provided by the user.
At the moment, if user fills with an option different than the default
empty, the produced nova.conf contains two [notifications] sections,
thus, ocurring in an erroneous configuration.

Change-Id: Iff44806bdd0f6078879c9872ba0531485f0ca4c7
Signed-off-by: Stamatis Katsaounis <skatsaounis@admin.grnet.gr>