Add/update region config in nova.conf
On multi-region deployments, Nova may talk to the wrong
neutron endpoint (from the wrong region) if the region
is unspecified.

On Rocky+ it will also require updating the
os_region_name config to region_name, as os_region_name
has been deprecated, otherwise Nova will talk to the wrong
placement endpoint as well.

This fix addresses the issue where nova-compute will not
register the node to the correct nova_api/placement
database, and will also not be able to complete live-migrations.

Given that the template for the [placement] section is
applied to every release, it is included both old and
new config options.

Change-Id: I9500ba400d55e6f1bc11f2ba05b25b4714cda578
Closes-bug: #1903210