Allow neutron-plugin sub to publish nothing
A change landed to charm helpers to mark the
SubordinateConfigContext as incomplete if the subordinate supplied
no data *1. But the neutron-plugin subordinate may legitimately
supply no data if no special config is needed. This restores the
previous behaviour of marking the subordinate context for the
neutron-plugin relation as complete even if not data was supplied.

*1 https://github.com/juju/charm-helpers/pull/519

Change-Id: I34fa2d39171132e4fe7d0b7e5fd29162161a5060
Closes-Bug: #1912187