Remove python-oslo.privsep on upgrades to >=rocky
Ubuntu OpenStack Rocky migrated from python2 to python3, the package
python-oslo.privsep provides the privsep-helper program as an
alternative, when having it installed next to python3-oslo.privsep the
first installed will stay as a default.

This change add python-oslo.privsep to the list of packages that need to
be purged on OpenStack>=rocky.

Change-Id: Ib64dfa30b6705a45f82051dfc3634740a5f3a661
Closes-Bug: #1822763