Update nova.conf for OpenStack Yoga
This merge request focuses on fixing the VNC config for OpenStack Yoga
on the Focal and Jammy series. Originally, the Yoga version of this
charm was using a template nova.conf file from Train which did not use
the new [vnc] required for Yoga. Train had the VNC config in the
[DEFAULT] section, which is depreciated in Yoga.

Itemized changes to charm below:

 * Create templates/yoga
 * Move VNC config to [vnc] section in nova.conf
 * Rename VNC config keys to follow the Yoga specification.

Closes-Bug: #1974082
Closes-Bug: #1734683

Change-Id: Ic100528f9f38bbc0c83e4f563166113024e3db59